requirements.txt
certifi~=2022.9.24
charset-normalizer==2.1.1
idna==3.4
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
APScheduler~=3.9.1
pip~=22.3
attrs~=22.1.0
wheel~=0.37.1
click~=8.1.3
six~=1.15.0
setuptools~=58.0.4
future~=0.18.2
aiohttp~=3.8.3